+++ src/main/kotlin/com/github/xepozz/testo/tests/TestoConsoleProperties.kt
@@ -97,7 +97,6 @@ class TestoConsoleProperties(
             testFrameworkName,
             consoleProperties,
             channelStore,
-            levelFilter,
             statusStore,
             runTimings,
             targetStore,